+++ src/main/java/com/example/mathsgenealogyapi/dissertation/DissertationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DissertationRepository extends JpaRepository<Dissertation, Long> {
+public interface DissertationRepository extends JpaRepository<Dissertation, Integer> {
 }
